== Happy Halloween == =General: *Merged missing changes from 3ceam rev. 778
*Please don't forget to add the upgrade_15-3athena_svn314.sql file!

*Changed default PACKETVER to 20180919
-The 2018-09-19aRagexe client is stable.
-This client is recommended if you want to use pet autofeeding.

*clif_parse_OpenVending will now fail with an empty item list.

*Added some missing packets for 2018 clients.

*Updated new char creation.
-The race field is supposed to be an int value, so there is no unknown byte in this packet. ^^

*Implemented Pet Autofeeding.
-We will need this soon to get the Pet Evolution System working.

*Fixed a issue where NPC chat rooms would fail to appear in certain conditions. Mainly with job quests.

+++ src/char/char.c
@@ -1354,7 +1354,7 @@ int check_char_name(char * name)
 //-----------------------------------
 #if PACKETVER >= 20120307
 #if PACKETVER >= 20151001
-int make_new_char(struct char_session_data* sd, char* name_, int slot, int hair_color, int hair_style, short race) {
+int make_new_char(struct char_session_data* sd, char* name_, int slot, int hair_color, int hair_style, int race) {
 #else
 int make_new_char(struct char_session_data* sd, char* name_, int slot, int hair_color, int hair_style) {
 #endif
@@ -1381,7 +1381,7 @@ int make_new_char(struct char_session_data* sd, char* name_, int str, int agi, i
 	// Race values are acturally sent by the client as the job ID the new character would start on.
 	// But to be safe, its best to have the server read what race was selected and then set the
 	// starting job itself rather then setting it to the value the client sent.
-	if ( race != RACE_HUMAN && race != RACE_DORAM ) {
+	if (race != RACE_HUMAN && race != RACE_DORAM ) {
 		ShowWarning("make_new_char: Detected character creation packet with invalid race type on account: %d.\n", sd->account_id);
 		return -2;
 	}
@@ -1436,7 +1436,7 @@ int make_new_char(struct char_session_data* sd, char* name_, int str, int agi, i
 	// Race selection from 0xa39 packet.
 	// 0 = Human
 	// 4218 = Doram
-	if ( race == RACE_HUMAN || ALLOW_OTHER_RACES == 0)
+	if (race == RACE_HUMAN || ALLOW_OTHER_RACES == 0)
 	{	// Human - Defaults
 		// Job = Novice
 		// Starting HP/SP = 40/11
